refactor: split rest page into components
